SmartMeeting/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
discord.py>=2.0.0
python-dotenv>=0.19.0
pytest>=6.2.5
pytest-cov>=2.12.0
requests>=2.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0